- port from opensips, r4550, Credits goes to Bogdan - fixed strange 64bits warning
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@4607 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/xcap_client/xcap_functions.c
@@ -560,7 +560,7 @@ char* send_http_get(char* path, unsigned int xcap_port, char* match_etag,
 	curl_easy_setopt(curl_handle, CURLOPT_WRITEHEADER, &etag);
 
 	if(match_header)
-		curl_easy_setopt(curl_handle, CURLOPT_HEADER, match_header);
+		curl_easy_setopt(curl_handle, CURLOPT_HEADER, (long)match_header);
 
 	/* non-2xx => error */
 	curl_easy_setopt(curl_handle, CURLOPT_FAILONERROR, 1);
